chore(project): move logger to aurelia-orm.js
