+++ src/aurelia-orm.js
@@ -32,3 +32,7 @@ export function configure(aurelia, configCallback) {
 }
 
 export {EntityManager, HasAssociationValidationRule, ValidationGroup};
+
+import {getLogger} from 'aurelia-logging';
+const logger = getLogger('aurelia-orm');
+export {logger};
